cli: protect against API nils
The newer API has more nested fields and they may be nil. The CLI
tooling needs to check this more rigourously.

+++ cilium/cmd/service_delete.go
@@ -36,6 +36,11 @@ var serviceDeleteCmd = &cobra.Command{
 			}
 
 			for _, svc := range list {
+				if svc.Status == nil || svc.Status.Realized == nil {
+					log.Error("Skipping service due to empty state")
+					continue
+				}
+
 				if err := client.DeleteServiceID(svc.Status.Realized.ID); err != nil {
 					log.WithError(err).WithField(logfields.ServiceID, svc.Status.Realized.ID).Error("Cannot delete service")
 				}
